remove user from household

+++ app/src/main/java/com/timgortworst/roomy/domain/usecase/SetupUseCase.kt
@@ -2,13 +2,13 @@ package com.timgortworst.roomy.domain.usecase
 
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
+import com.google.firebase.firestore.FirebaseFirestoreException
 import com.timgortworst.roomy.R
 import com.timgortworst.roomy.data.repository.HouseholdRepository
 import com.timgortworst.roomy.data.repository.IdProvider
 import com.timgortworst.roomy.data.repository.TaskRepository
 import com.timgortworst.roomy.data.repository.UserRepository
 import com.timgortworst.roomy.domain.model.ResponseState
-import com.timgortworst.roomy.domain.model.Role
 
 class SetupUseCase(
     private val householdRepository: HouseholdRepository,
@@ -17,53 +17,51 @@ class SetupUseCase(
     private val idProvider: IdProvider
 ) {
     suspend fun switchHousehold(newId: String) {
-        val oldId= idProvider.getHouseholdId()
+        val oldId = idProvider.fetchHouseholdId()
         val currentUserId = FirebaseAuth.getInstance().currentUser?.uid ?: return
 
-        // update user with new household ID and role
-        userRepository.updateUser(
-            userId = currentUserId,
-            householdId = newId,
-            role = Role.NORMAL.name
-        )
+        try {
+            // remove old tasks assigned to user
+            taskRepository.deleteTasks(
+                taskRepository.getTasksForUser(currentUserId)
+            )
 
-        // remove old tasks assigned to user
-        val tasks= taskRepository.getTasksForUser(currentUserId)
-        taskRepository.deleteTasks(tasks)
-
-        // delete old household if no user is assigned to it
-        userRepository.getAllUsersForHousehold(oldId)?.let {
-            if (it.isEmpty()) {
-                householdRepository.deleteHousehold(oldId)
+            // delete old household if no other user is left to assigned to it
+            userRepository.getAllUsersForHousehold(oldId).let {
+                if (it.size <= 1) householdRepository.deleteHousehold(oldId)
             }
+
+            // update user with new household ID and role
+            userRepository.updateUser(
+                userId = currentUserId,
+                householdId = newId,
+                isAdmin = false
+            )
+        } catch (e: FirebaseFirestoreException) {
+            // todo handle errors
         }
     }
 
-    suspend fun currentHouseholdIdForCurrentUser() = idProvider.getHouseholdId()
+    suspend fun currentHouseholdIdForCurrentUser() = idProvider.fetchHouseholdId()
 
     suspend fun isIdSimilarToActiveId(referredHouseholdId: String): Boolean {
-        return referredHouseholdId == idProvider.getHouseholdId()
+        return referredHouseholdId == idProvider.fetchHouseholdId()
     }
 
     suspend fun handleLoginResult(
         fbUser: FirebaseUser?,
         newUser: Boolean
     ): ResponseState {
-        val user = fbUser ?: return ResponseState.Error(R.string.error_generic)
+        val fireBaseUser = fbUser ?: return ResponseState.Error(R.string.error_generic)
 
         if (newUser) {
-            val householdId = createNewHousehold() ?: run {
+            try {
+                val householdId = householdRepository.createHousehold()
+                userRepository.createUser(householdId, fireBaseUser)
+            } catch (e: FirebaseFirestoreException) {
                 return ResponseState.Error(R.string.error_generic)
             }
-            createNewUser(householdId, user)
         }
         return ResponseState.Success(fbUser.displayName)
     }
-
-    private suspend fun createNewHousehold() = householdRepository.createHousehold()
-
-    private suspend fun createNewUser(
-        householdId: String,
-        fireBaseUser: FirebaseUser
-    ) = userRepository.createUser(householdId, fireBaseUser)
 }
